# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A modular and scalable Go REST API template implementing Clean Architecture.",
        "title": "Go Clean Template API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/sessions/revoke-all": {
            "post": {
                "description": "Revoke all user sessions except the current one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Revoke all other sessions",
                "parameters": [
                    {
                        "description": "Revoke sessions request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RevokeSessionsIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.RevokeSessionsIn": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "gct_internal_controller_restapi_response.ErrorDetail": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "DATA"
                },
                "code": {
                    "type": "string",
                    "example": "RESOURCE_NOT_FOUND"
                },
                "details": {
                    "type": "string",
                    "example": "The user with the ID '12345' does not exist in our records."
                },
                "message": {
                    "type": "string",
                    "example": "The requested resource was not found."
                },
                "method": {
                    "type": "string",
                    "example": "GET"
                },
                "path": {
                    "type": "string",
                    "example": "/api/v1/users/12345"
                },
                "retry_after": {
                    "description": "seconds",
                    "type": "integer",
                    "example": 5
                },
                "retryable": {
                    "type": "boolean",
                    "example": false
                },
                "severity": {
                    "description": "Enhanced fields",
                    "type": "string",
                    "example": "MEDIUM"
                },
                "suggestion": {
                    "type": "string",
                    "example": "Please check our documentation."
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-08T12:30:45Z"
                }
            }
        },
        "gct_internal_controller_restapi_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "documentation_url": {
                    "type": "string",
                    "example": "https://developer.mozilla.org/en-US/docs/Web/HTTP/Status"
                },
                "error": {
                    "$ref": "#/definitions/gct_internal_controller_restapi_response.ErrorDetail"
                },
                "requestId": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-g1h2-i3j4k5l6m7n8"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "gct_internal_controller_restapi_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "meta": {},
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}